Allow streaming to be paused repeatedly
This is required when reading from rate-limited mongodb servers.
A clear example is Azure CosmosDB that is rate limited by the
provisioned RUs. If not pausing, the read cursor is killed and the
stream is closed, throwing a mongo exception and halting the whole
process.
To control this set 2 environment variables:
1. PAUSE_EVERY (number of documents)
2. PAUSE_FOR (in seconds)